// frontend/src/modules/client/routes/enums/route.enum.ts
export enum ClientRouteEnum {
  ROOT = '/client',
  SIGN_IN = '/client/signin',
  SIGN_UP = '/client/signup',
  SIGN_UP_LAST_STEP = '/client/signup-last-step',
  PASSWORD_RECOVERY = '/client/forgot-password',
  BOOKINGS = '/client/bookings',
  REBOOK = '/client/rebook',
  REBOOK_APPOINTMENT = '/client/bookings/rebook',
  CREATE_APPOINTMENT = '/client/bookings/new',
  HELP = '/client/help',
}