// frontend/src/modules/public/routes/enums/route.enum.ts
export enum PublicRouteEnum {
  ROOT = '',
  LANDING = '/landing',
  BOOK_APPOINTMENT = '/new-appointment',
  EMAIL_VERIFICATION = '/apply-verification-code/:id',
  APPLY_VERIFICATION_CODE = '/apply-verification-code/:id/:code',
  PRIVACY_POLICY = '/privacy-policy',
  TERMS_AND_CONDITIONS = '/terms-and-conditions',
  CANCELLATION_POLICY = '/cancellation-policy',
}